Improve progress percentage
Currently the progress is shown as 20/100% which is not ideal as percentages already mean xy/100. This commit changes that format to xy.zw%

where zw is two digits of precision. this is done to give the user more frequent feedback on progress as previously the number was rounded and would update only once it reached the next int